extract PendingRequest.java
