+++ src/main/java/org/littleshoot/proxy/impl/ServerConnectionPool.java
@@ -82,29 +82,4 @@ public interface ServerConnectionPool {
 
   /** Returns the maximum total number of connections allowed in the pool. */
   int getMaxConnections();
-
-  /** Tracks a pending request and its associated client connection for HTTP pipelining. */
-  class PendingRequest {
-    private final ClientToProxyConnection clientConnection;
-    private final HttpRequest request;
-    private final long timestamp;
-
-    public PendingRequest(ClientToProxyConnection clientConnection, HttpRequest request) {
-      this.clientConnection = clientConnection;
-      this.request = request;
-      this.timestamp = System.currentTimeMillis();
-    }
-
-    public ClientToProxyConnection getClientConnection() {
-      return clientConnection;
-    }
-
-    public HttpRequest getRequest() {
-      return request;
-    }
-
-    public long getTimestamp() {
-      return timestamp;
-    }
-  }
 }
